+++ Runtime/Scripts/Core/WTResources.cs
@@ -4,7 +4,7 @@ namespace WiseTrack.Core
 {
     public class WTResources
     {
-        public static readonly string SdkVersion = "2.0.2";
+        public static readonly string SdkVersion = "2.0.3";
         public static readonly WTLogLevel DefaultLogLevel = WTLogLevel.Debug;
         public static WTSDKEnvironment DefaultSdkEnv = WTSDKEnvironment.Production;
     }