+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [2.0.3]
+
+### Enhanced
+
+- Updated WiseTrack SDK to version 2.2.2
+- Enhanced event parameter handling and validations
+- Updated Android plugin integration
+
 ## [2.0.2]
 
 ### Chore